+++ package.json
@@ -7,6 +7,7 @@
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^0.24.0",
+    "bootstrap": "^5.1.3",
     "history": "^5.2.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
